+++ crates/miden-standards/src/account/auth/mod.rs
@@ -2,7 +2,7 @@ mod approver;
 pub use approver::{Approver, ApproverSet};
 
 mod fee;
-pub use fee::FeePaymentInfo;
+pub use fee::FeeConversionInfo;
 
 mod no_auth;
 pub use no_auth::NoAuth;